fix(debian): add support for arm64 architecture
Instead of having a series of very complicated if-then-else statements
inside of the definitions, this uses `with_first_found` to define the
architectures. In general, it makes things a lot easier for defining the
code and leads to other nice things like including SHA sums for multiple
different distributables.

Also, bumped to version of git-delta to 0.6.0.

All of this was needed because I now have a 64 bit ARM board running a
64 bit OS (NVIDIA Jetson Nano).

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>