fix(delta): make command line invoke normal delta command
Previously this tried to do some various forms of styling on the output
to delta - but the problem is that those command line flags are out of
date and trying to port them over actually makes things really ugly now.
The defaults to delta are good enough for us and this make that so.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>